requirements.txt
boto3==1.24.8
PyYAML==6.0
